+++ src/components/datatable/EarthEngineColumns.js
@@ -3,18 +3,18 @@ import PropTypes from 'prop-types';
 import { Column } from 'react-virtualized';
 import ColumnHeader from './ColumnHeader';
 import { numberPrecision } from '../../util/numbers';
-import { getPropName, getPrecision } from '../../util/earthEngine';
+import { hasClasses, getPrecision } from '../../util/earthEngine';
 
-const EarthEngineColumns = ({ classes, aggregationType, legend, data }) => {
+const EarthEngineColumns = ({ aggregationType, legend, data }) => {
     const { title, items } = legend;
 
-    if (classes && items) {
+    if (hasClasses(aggregationType) && items) {
         const valueFormat = numberPrecision(2);
 
         return items.map(({ id, name }) => (
             <Column
                 key={id}
-                dataKey={name}
+                dataKey={String(id)}
                 label={name}
                 width={100}
                 className="right"
@@ -28,15 +28,15 @@ const EarthEngineColumns = ({ classes, aggregationType, legend, data }) => {
         ));
     } else if (Array.isArray(aggregationType) && aggregationType.length) {
         return aggregationType.map(type => {
-            const propName = getPropName(type, title);
-            const precision = getPrecision(data.map(d => d[propName]));
+            const label = `${type} ${title}`.toUpperCase(); // Already translated
+            const precision = getPrecision(data.map(d => d[type]));
             const valueFormat = numberPrecision(precision);
 
             return (
                 <Column
                     key={type}
-                    dataKey={propName}
-                    label={`${type} ${title}`}
+                    dataKey={type}
+                    label={label}
                     width={100}
                     className="right"
                     headerRenderer={props => (
@@ -58,6 +58,7 @@ EarthEngineColumns.propTypes = {
     aggregationType: PropTypes.array,
     legend: PropTypes.object,
     data: PropTypes.array,
+    aggregations: PropTypes.object,
 };
 
 export default EarthEngineColumns;