+++ src/components/map/layers/earthEngine/EarthEngineLayer.js
@@ -2,8 +2,9 @@ import React from 'react';
 import Layer from '../Layer';
 import LayerLoading from '../LayerLoading';
 import EarthEnginePopup from './EarthEnginePopup';
+import Alert from '../Alert';
 import { apiFetch } from '../../../../util/api';
-import { getPropName, hasClasses } from '../../../../util/earthEngine';
+import { filterData } from '../../../../util/filter';
 import { EARTH_ENGINE_LAYER } from '../../../../constants/layers';
 
 export default class EarthEngineLayer extends Layer {
@@ -11,6 +12,7 @@ export default class EarthEngineLayer extends Layer {
         isLoading: true,
         popup: null,
         aggregations: null,
+        error: null,
     };
 
     componentDidUpdate(prev) {
@@ -24,14 +26,24 @@ export default class EarthEngineLayer extends Layer {
                     lng: coordinate[0],
                     lat: coordinate[1],
                 });
-            } catch (err) {
-                // eslint-disable-next-line
-                console.error(
-                    'Google Earth Engine failed. Is the service configured for this DHIS2 instance?'
-                );
+            } catch (error) {
+                this.setState({
+                    error:
+                        'Google Earth Engine failed. Is the service configured for this DHIS2 instance?',
+                });
             }
         }
     }
+
+    updateLayer = filterChange => {
+        if (filterChange) {
+            this.applyFilter();
+        } else {
+            this.removeLayer();
+            this.createLayer(true);
+            this.setLayerOrder();
+        }
+    };
 
     createLayer(isUpdate) {
         const {
@@ -79,7 +91,7 @@ export default class EarthEngineLayer extends Layer {
             name,
             unit,
             value,
-            legend: legend ? legend.items : null,
+            legend: legend.items,
             resolution,
             projection,
             data,
@@ -106,50 +118,70 @@ export default class EarthEngineLayer extends Layer {
 
         config.accessToken = apiFetch('/tokens/google'); // returns promise
 
-        this.layer = map.createLayer(config);
-        map.addLayer(this.layer);
+        try {
+            this.layer = map.createLayer(config);
+            map.addLayer(this.layer);
+        } catch (error) {
+            this.onError(error);
+        }
 
         if (aggregate) {
             this.layer
                 .aggregate(aggregationType)
-                .then(this.addAggregationValues.bind(this));
+                .then(this.addAggregationValues.bind(this))
+                .catch(this.onError.bind(this));
         }
 
         this.fitBoundsOnce();
     }
 
     addAggregationValues(aggregations) {
-        const { aggregationType, data, legend } = this.props;
-        const { title = '', items } = legend;
-        const classes = hasClasses(aggregationType);
+        const { id, data, setAggregations } = this.props;
 
-        // Make aggregations available for data table/download
-        data.forEach(f => {
-            const values = aggregations[f.id];
+        // Make aggregations available for data table and download
+        // setAggregations is not available in map plugin
+        if (setAggregations) {
+            setAggregations({ [id]: aggregations });
+        }
 
-            if (values) {
-                if (classes && items) {
-                    items.forEach(({ id, name }) => {
-                        f.properties[name] = values[id];
-                    });
-                } else {
-                    Object.keys(values).forEach(
-                        key =>
-                            (f.properties[getPropName(key, title)] =
-                                values[key])
-                    );
-                }
-            }
-            f.properties.type = f.geometry.type;
+        // Make aggregations available for filtering and popup
+        this.setState({
+            data: data.map(f => ({
+                ...f,
+                properties: {
+                    ...f.properties,
+                    ...aggregations[f.id],
+                },
+            })),
+            aggregations,
         });
+    }
 
-        // Make aggregations available for popup
-        this.setState({ aggregations });
+    applyFilter() {
+        const { data, dataFilters } = this.props;
+
+        const filteredData = filterData(
+            this.state.data || data,
+            dataFilters
+        ).map(f => f.id);
+
+        if (this.layer && this.layer.filter) {
+            this.layer.filter(filteredData);
+        }
     }
 
     render() {
         const { legend, aggregationType } = this.props;
-        const { isLoading, popup, aggregations } = this.state;
+        const { isLoading, popup, aggregations, error } = this.state;
+
+        if (error) {
+            return (
+                <Alert
+                    message={error}
+                    onHidden={() => this.setState({ error: null })}
+                />
+            );
+        }
 
         if (isLoading) {
             return <LayerLoading />;
@@ -171,6 +203,10 @@ export default class EarthEngineLayer extends Layer {
     }
 
     onLoad() {
-        this.setState({ isLoading: false });
+        this.setState({ isLoading: false, popup: null });
+    }
+
+    onError(error) {
+        this.setState({ error, isLoading: false });
     }
 }