+++ src/epics/maps.js
@@ -5,7 +5,7 @@ import * as types from '../constants/actionTypes';
 import { errorActionCreator } from '../actions/helpers';
 import { mapRequest } from '../util/requests';
 import { addOrgUnitPaths } from '../util/helpers';
-import { setMap } from '../actions/map';
+import { setMap, setMapProps } from '../actions/map';
 import { loadLayer } from '../actions/layers';
 
 // Load one favorite
@@ -22,4 +22,18 @@ export const loadFavorite = action$ =>
             ...addOrgUnitPaths(config.mapViews).map(loadLayer),
         ]);
 
-export default combineEpics(loadFavorite);
+// Update favorite except layers on map save
+export const updateFavorite = action$ =>
+    action$
+        .ofType(types.FAVORITE_UPDATE)
+        .concatMap(action =>
+            mapRequest(action.id).catch(
+                errorActionCreator(types.FAVORITE_UPDATE_ERROR)
+            )
+        )
+        .mergeMap(config => {
+            delete config.mapViews;
+            return [setMapProps(config)];
+        });
+
+export default combineEpics(loadFavorite, updateFavorite);